requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.36
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
bcrypt==4.0.1
python-multipart==0.0.9
requests

